fixed int division

+++ src/auto-evo/simulation/niche/HeterotrophicNiche.cs
@@ -1,6 +1,5 @@
 ﻿namespace Thrive.src.auto_evo.simulation.niche
 {
-    using System;
     using System.Linq;
 
     public class HeterotrophicNiche : Niche
@@ -37,7 +36,7 @@
             var preySize = microbeSpecies.Organelles.Organelles.Sum(organelle => organelle.Definition.HexCount);
 
             // It's great if you can engulf this prey, but only if you can catch it
-            var engulfScore = predatorSize / preySize > Constants.ENGULF_SIZE_RATIO_REQ ?
+            var engulfScore = (float)predatorSize / (float)preySize > Constants.ENGULF_SIZE_RATIO_REQ ?
                 Constants.AUTO_EVO_ENGULF_PREDATION_SCORE : 0.0f;
             engulfScore *= predatorSpeed > preySpeed ? 1.0f : 0.1f;
